+++ Gruntfile.js
@@ -10,6 +10,17 @@
  */
 module.exports = function(grunt) {
 
+  //--> Utilities:
+  const INDEX_TAG = "index";
+  const DECLATE_TAG = /declare /g;
+  const IMPORT_STATEMENT = /import.*;/g;
+  const TRIM_BLANK_LINES = /^\s*[\r\n]/gm;
+  const PLACEHOLDER = /\/\*PLACEHOLDER\*\//g;
+  const EMPTY_STRING = "";
+
+  //--> Stores types during build process:
+  let typesContent = EMPTY_STRING;
+
   //--> Grunt config initialization:
   grunt.initConfig({});
 
@@ -47,13 +58,47 @@ module.exports = function(grunt) {
     }
   });
 
-  //--> Copies all ".js" files from the "src" to the "lib" folder:
+  //--> Copy tasks:
   grunt.config("copy", {
-    files: {
-      cwd: "src",
-      src: "**/*.js",
-      dest: "lib",
+    //--> Copies all ".js" files from the "src" to the "lib" folder:
+    main: {
+        cwd: "src",
+        src: "**/*.js",
+        dest: "lib",
+        expand: true
+    },
+    //--> Initializes all type files:
+    initTypes: {
+      cwd: "types/temp",
+      src: "**/*.ts",
+      dest: "types/temp2",
       expand: true,
+      flatten: true,
+      filter: "isFile",
+      options: {
+        process: function (content, srcpath) {
+          if(srcpath.indexOf(INDEX_TAG) === -1) {
+            let result =  content.replace(DECLATE_TAG, EMPTY_STRING);
+            result = result.replace(IMPORT_STATEMENT, EMPTY_STRING);
+            result = result.replace(TRIM_BLANK_LINES, EMPTY_STRING);
+            typesContent += result;
+          }
+          return content;
+        }
+      }
+    },
+    //--> Builds final type file:
+    buildTypesFile: {
+      cwd: "utils/types",
+      src: "**/*.ts",
+      dest: "types",
+      expand: true,
+      options: {
+        process: function (content, srcpath) {
+          content = content.replace(PLACEHOLDER, typesContent);
+          return content;
+        }
+      }
     }
   });
   
@@ -70,6 +115,11 @@ module.exports = function(grunt) {
         exclude: "**/*Test*.ts"
       }
     }
+  });
+
+  //--> Removes temp folders:
+  grunt.config("clean", {
+    build: ["types/temp", "types/temp2"]
   });
 
   //--> Runs Unit Testing:
@@ -91,13 +141,18 @@ module.exports = function(grunt) {
 
   //--> Imports the task used for running tests:
   grunt.loadNpmTasks('grunt-mocha-test');
+    
+  //--> Imports the task used for deleting temp files:
+  grunt.loadNpmTasks('grunt-contrib-clean');
   
   /*!
    * JEC Tiger Tasks:
    */
 
   //--> Task: builds the "jec-tiger" project:
-  grunt.registerTask("build", ["ts:build", "copy"]);
+  grunt.registerTask("build", ["ts:build", "copy:main",
+                               "copy:initTypes", "copy:buildTypesFile",
+                               "clean:build"]);
 
   //--> Task: builds the "jec-tiger" API documentation:
   grunt.registerTask("doc", ["typedoc"]);